+++ src/jetstream/custom_error/custom_error.go
@@ -1,4 +1,4 @@
-package errorz
+package custom_error
 
 import "errors"
