+++ src/jetstream/plugins/kubernetes/main.go
@@ -12,7 +12,7 @@ import (
 	"errors"
 
 	"github.com/cloudfoundry-incubator/stratos/src/jetstream/api"
-	"github.com/cloudfoundry-incubator/stratos/src/jetstream/errorz"
+	"github.com/cloudfoundry-incubator/stratos/src/jetstream/custom_error"
 	"github.com/labstack/echo/v4"
 	log "github.com/sirupsen/logrus"
 
@@ -289,7 +289,7 @@ func (c *KubernetesSpecification) RequiresCert(ec echo.Context) error {
 		Message  string
 	}
 	if err != nil {
-		if strings.Contains(err.Error(), errorz.ERR_X509_CERTIFICATE) {
+		if strings.Contains(err.Error(), custom_error.ERR_X509_CERTIFICATE) {
 			response.Status = http.StatusOK
 			response.Required = true
 		} else {
