+++ src/jetstream/repository/goose-db-version/pgsql_goose_db_version_test.go
@@ -10,7 +10,7 @@ import (
 	. "github.com/smartystreets/goconvey/convey"
 
 	"github.com/cloudfoundry-incubator/stratos/src/jetstream/api"
-	"github.com/cloudfoundry-incubator/stratos/src/jetstream/errorz"
+	"github.com/cloudfoundry-incubator/stratos/src/jetstream/custom_error"
 )
 
 func TestPgSQLGooseDB(t *testing.T) {
@@ -72,7 +72,7 @@ func TestPgSQLGooseDB(t *testing.T) {
 			Convey("there should be an error", func() {
 				repository, _ := NewPostgresGooseDBVersionRepository(db)
 				_, err := repository.GetCurrentVersion()
-				So(err, ShouldResemble, errorz.ErrNoDatabaseVersionsFound)
+				So(err, ShouldResemble, custom_error.ErrNoDatabaseVersionsFound)
 
 				dberr := mock.ExpectationsWereMet()
 				So(dberr, ShouldBeNil)
